backend/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
shopifyapi>=9.0.0
ebaysdk>=2.2.0
pytrends>=4.8.0
jinja2>=3.0.0
pyyaml>=5.4.0
requests>=2.26.0
